Dupe function fixed
Fixed the bug wahoo! So now prior to writing to file, it scans the existing file. if the file is empty or doesn't exist, it'll write the current inventory (the one it just scanned) into it (as everything it adds will be new. If the file is already occupied, it will check the new inventory against existing entries and only add to file if it doesn't previously exist